sec(1000-agent): fix M-1/L-1/L-2/L-3/L-4 + 3 gas packings (redeploy batch)
Implements all five findings from the 2026-07-22 1000-agent per-contract audit
(docs/AUDIT_2026_07_22_1000AGENT.md). All are Low/Medium and fail-closed; the
on-chain read-back (block 25593551, publicnode+drpc) confirmed zero stranded
value and that the vulnerable setters are absent from live bytecode, so this is
a pre-emptive redeploy batch with no recovery migration required.

M-1 (Med) SwapFeeRouter sub-18-decimal fee stranding — two causes:
  - lib/SwapFeeRouterConvertLib.sol: flat 1e18 raw floor -> per-token
    minTokenFeeForConversion() = 10**decimals() (no-op at 18dp; low-level
    staticcall so non-standard tokens degrade to the 18dp fallback, never looser)
  - SwapFeeRouter.sol: withdraw/sweep gated on pair EXISTENCE over-shot; an
    unfunded pair left all three exits closed. Now requires non-zero reserves
    via _hasConvertibleWethPair. Corrected the misleading in-code operator note.

L-1 (Low) setSequencerFeed missing mainnet guard — add
  `if (block.chainid == 1) revert SequencerFeedNotOnMainnet()` on SwapFeeRouter
  and TegridyNFTLending, mirroring TegridyLending.sol:840 / SequencerCheck.sol:163.

L-2 (Low) principal min/max window lacked the cross-guard — added at BOTH
  propose-time (TegridyLendingAdmin) and apply-time (TegridyLending); only the
  apply-time re-check blocks the parallel-timelock inversion. Mirrors the APR /
  duration siblings. (TegridyLending is undeployed.)

L-3 (Low) FoT conversion floor sized on gross input stranded high-tax tokens —
  price component now rescaled by the first-hop pair's measured balanceOf delta;
  caller floor + absolute anti-drain floor never rescaled; MAX_FOT_TAX_BPS=3000
  bounds the haircut (reverts FoTHaircutTooLarge beyond it).

L-4 (Low) TegridyTWAP deviation-baseline ratchet — lastSpot seeded from the
  interval TWAP instead of raw instantaneous spot; branch conditions unchanged.

Gas (storage-layout, fresh-deploy only — measured, not estimated):
  KEPT: GaugeController VoteAllocation 2->1 slot (-13 B, -65,886 gas/vote);
    MemeBountyBoard Bounty 9->4 slots (+257 B, -32% createBounty);
    TegridyNFTLending/TegridyLending Offer.active reorder.
  REVERTED: RevenueDistributor Epoch (+548 B) and VoteIncentives EpochInfo
    (+110 B) — traded scarce EIP-170 headroom for per-epoch gas; not worth it.

Tests: 38 new, all mutation-verified (each fix reverted in place -> its test
fails). Adds contracts/test/Audit_1000Agent_{SFR,Lending,TWAP,FeeConservation}.t.sol.
The FeeConservation harness asserts REACHABILITY (every material balance has an
exit), which conservation alone would miss. Placed at test/ root because the CI
matrix never descends into test/invariants/. Full suite: 1903 pass, 0 fail;
CI size-budget gate green.

+++ contracts/src/GaugeController.sol
@@ -205,9 +205,26 @@ contract GaugeController is OwnableNoRenounce, ReentrancyGuard, Pausable, Timelo
     /// @notice Stores each tokenId's vote allocations for the epoch they voted in
     mapping(uint256 => VoteAllocation[]) internal _tokenVotes;
 
+    /// @dev GAS (1000-agent audit 2026-07-22, HIGH): packed 2 storage slots → 1.
+    ///      `_tokenVotes` is a STORAGE array and both vote paths `push` one entry per
+    ///      gauge, so the second slot cost a full 20,000-gas cold SSTORE per gauge —
+    ///      ~160,000 gas on a maximum 8-gauge vote.
+    ///
+    ///      `weight` is a BPS share, hard-capped at `MAX_WEIGHT_PER_GAUGE_BPS` (5000)
+    ///      and required to sum to `BPS` (10000). Both writers — `vote()` (:447) and
+    ///      `revealVote()` (:733) — enforce that cap on every element BEFORE the push,
+    ///      so the `uint96` narrowing cannot truncate: the domain is [1, 5000] and
+    ///      `uint96` holds ~7.9e28. `uint96` (12 bytes) is chosen over a tighter
+    ///      `uint16` because it exactly fills the 12 bytes left by `address`, giving
+    ///      free headroom without costing a second slot.
+    ///
+    ///      STORAGE-LAYOUT CHANGE — fresh deploy only; a live instance cannot be
+    ///      repacked in place. Also narrows the `getTokenVotes` return tuple from
+    ///      `(address,uint256)` to `(address,uint96)`; that view has no in-repo
+    ///      consumers, but regenerate the frontend ABI on deploy.
     struct VoteAllocation {
         address gauge;
-        uint256 weight;
+        uint96 weight;
     }
 
     // ─── Commit-Reveal Voting State ─────────────────────────────────
@@ -473,7 +490,7 @@ contract GaugeController is OwnableNoRenounce, ReentrancyGuard, Pausable, Timelo
             uint256 allocatedPower = (votingPower * weights[i]) / BPS;
             gaugeWeightByEpoch[epoch][gauges[i]] += allocatedPower;
             totalWeightByEpoch[epoch] += allocatedPower;
-            _tokenVotes[tokenId].push(VoteAllocation({gauge: gauges[i], weight: weights[i]}));
+            _tokenVotes[tokenId].push(VoteAllocation({gauge: gauges[i], weight: uint96(weights[i])}));
             // AUDIT FIX FRESH-2026: F-17-4 — `_updateEpochTop` removed (write-only
             // dead state; not read anywhere after V3-GOV-03 + V3-GOV-06 rewrote
             // `_getRelativeWeightAt` to use only `totalWeightByEpoch`).
@@ -764,7 +781,7 @@ contract GaugeController is OwnableNoRenounce, ReentrancyGuard, Pausable, Timelo
             uint256 allocatedPower = (votingPower * weights[i]) / BPS;
             gaugeWeightByEpoch[epoch][gauges[i]] += allocatedPower;
             totalWeightByEpoch[epoch] += allocatedPower;
-            _tokenVotes[tokenId].push(VoteAllocation({gauge: gauges[i], weight: weights[i]}));
+            _tokenVotes[tokenId].push(VoteAllocation({gauge: gauges[i], weight: uint96(weights[i])}));
             // AUDIT FIX FRESH-2026: F-17-4 — `_updateEpochTop` removed (dead state).
         }
